requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
alabaster==1.0.0
    # via sphinx
anyio==4.8.0
    # via httpx
async-timeout==5.0.1
    # via redis
babel==2.16.0
    # via sphinx
bibtexparser==1.4.3
    # via pyzotero
blurhash==1.1.4
    # via mastodon-py
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
decorator==5.1.1
    # via mastodon-py
docutils==0.21.2
    # via
    #   sphinx
    #   sphinx-rtd-theme
exceptiongroup==1.2.2
    # via anyio
feedparser==6.0.11
    # via pyzotero
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via pyzotero
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
imagesize==1.4.1
    # via sphinx
jinja2==3.1.5
    # via sphinx
markupsafe==3.0.2
    # via jinja2
mastodon-py==1.8.1
    # via -r requirements.in
packaging==24.2
    # via sphinx
pygments==2.19.1
    # via sphinx
pyparsing==3.2.1
    # via bibtexparser
python-dateutil==2.9.0.post0
    # via mastodon-py
python-dotenv==1.0.1
    # via -r requirements.in
python-magic==0.4.27
    # via mastodon-py
pytz==2024.2
    # via pyzotero
pyzotero==1.6.5
    # via -r requirements.in
redis==5.2.1
    # via -r requirements.in
requests==2.32.3
    # via
    #   -r requirements.in
    #   mastodon-py
    #   sphinx
sgmllib3k==1.0.0
    # via feedparser
six==1.17.0
    # via
    #   mastodon-py
    #   python-dateutil
sniffio==1.3.1
    # via anyio
snowballstemmer==2.2.0
    # via sphinx
sphinx==8.1.3
    # via
    #   sphinx-rtd-theme
    #   sphinxcontrib-jquery
sphinx-rtd-theme==3.0.2
    # via pyzotero
sphinxcontrib-applehelp==2.0.0
    # via sphinx
sphinxcontrib-devhelp==2.0.0
    # via sphinx
sphinxcontrib-htmlhelp==2.1.0
    # via sphinx
sphinxcontrib-jquery==4.1
    # via sphinx-rtd-theme
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==2.0.0
    # via sphinx
sphinxcontrib-serializinghtml==2.0.0
    # via sphinx
tomli==2.2.1
    # via sphinx
typing-extensions==4.12.2
    # via anyio
urllib3==2.3.0
    # via requests
